billing-service/requirements.txt
gunicorn==20.1.0
falcon==3.0.1
pact-python==2.1.1